A fair bit of work to be done on the site ...
See #16 for more info. Probably needs a different font and think about shrinking down the size for legibility ...

Right now it goes from small `--font-size` to massive `--font-size-p` and the baseline grid might not be suitable for `17px` or inbetween sizes as everything is divisible by `4` or `8`

I might even have to change the baseline grid.