Vimeo and javascript is too unpredictable
Just use regular links, or host the videos individually. I'm leaving this as it is (unfinished) and reverting to plain old links to Vimeo.

I can't get this damned thing to work, and Vimeo's caching of the video (even if you're assigning the correct `video_id`) is frustrating. I can't figure out how to render a new video reliably, and it shouldn't be this difficult.